feat: rename dateHelper and transfer two methods from component to this utils file
